requirements.txt
Flask
jsonpickle
jsonschema
boto3
wallet-py3k